Refactor, clean-up and consolidate tests
